+++ src/pages/SigningOfficialDarRequests.js
@@ -7,7 +7,7 @@ import {Collections} from '../libs/ajax';
 import { USER_ROLES } from '../libs/utils';
 import { DarCollectionTableColumnOptions, DarCollectionTable } from '../components/dar_collection_table/DarCollectionTable';
 import { consoleTypes } from '../components/dar_collection_table/DarCollectionTableCellData';
-import SigningOfficialDAAPopup from '../components/modals/SigningOfficialDAAPopup';
+import SigningOfficialDAAPopup from '../components/SigningOfficialDAAPopup';
 
 export default function SigningOfficialDarRequests() {
   const [collectionList, setCollectionList] = useState([]);
